Update Chapter01.md
Defined the terms Instance Type and Instance:

Instance Type; defines the capabilities of an image available to users for the creation of a virtual compute resource; includes capabilities such as CPU, storage, and memory.

Instance: is a virtual compute resource, in a known state such as running or suspended, that can be used like a physical server. NOTE: can be used to specify VM Instance or Container Instance.