Update Chapter01.md
Changed definition of Instance Type:
Instance type: specifies a set of virtualized hardware resources and capabilities used for the creation of a virtual compute on which a workload runs on; includes capability specifications such as CPU, storage, and memory.

It is more in line with multiple versions:
Google: A machine type is a set of virtualized hardware resources available to a virtual machine (VM) instance, including the system memory size, virtual CPU (vCPU) count, and persistent disk limits.
 
AWS: Instance Types. When you launch an instance, the instance type that you specify determines the hardware of the host computer used for your instance. ... Amazon EC2 dedicates some resources of the host computer, such as CPU, memory, and instance storage, to a particular instance.

OpenStack: Instance Type: Describes the parameters of the various virtual machine images that are available to users; includes parameters such as CPU, storage, and memory. Alternative term for flavor.