Disambiguate VNF vs CNF
Trying to disambiguate the definitions of VNFs and CNFs. Current definition mixes (at least) two different aspects: delivery as virtual appliance vs as container and whether or not the implementation adheres to cloud-native principles.
Resolving this in terms of VMs vs containers, a) for consistency with NF definition above and b) because this aspect actually impacts the following sections.