[RM Glossary] Fix some container related definitions (#785)
* Fix CNF definition

Comment on #595: The last two sentences reference "containers" but container is not previously mentioned. "... run on an application orchestration layer ..." -- do we need "application"?
Typo, last sentence, capitalize "t" in "the containers themselves ... "

* Add note on OCI definition to Container

as proposed in a comment to #595

* Move container definition from 11.1.1 to the note in 11.1.4

There should only be one definition for container

* Add definition for Container Runtime

as requested in comment to #595.

* Move definition of Container Engine

Move all container related definitions to 11.1.4

* Fixes on Kubernetes Cluster

as requested in comments to #595.

* Clarification to Kubernetes Service

based on a comment to #595, since the term network service is used differently by the Kubernetes Glossary.
Also fixing other links to glossaries.

* Resolve comments on several definitions

Resolve comments on CaaS, Caas Manager, Container Runtime, Kubernetes Cluster, Kubernetes Master.

* Note on TUG Whitepaper added and an editorial fix

Added that CNCF TUG Whitepaper is still a draft
Added "term" and """ to network service in definition of Kubernetes Service

* Fix definition of kubernetes cluster

Adapt the wording from Kubernetes glossary to make it more consistent, since in Kubernetes definition, the master is not called a node or a worker machine.