[RM Ch04] Update metrics for Storage (#963)
It looks like the write IOPS for bronze was quite out of line with the other metrics.

Has any thought been given to synthetic workload for the disk test?  Is it based on a single thread of I/O waiting for completion?  Or async, with multiple requests in flight?

Also, will there be latency requirements?