Add oauth1 as optional Keystone features
It was not listed in Keystone which defacto sets it as optional.
It may be considered as mandatory in a second step.

It should be noted that it's not working in CNTT RI and there is no
true plan to update CNTT RI from the time writing.

RA1 and RC are updated in one change according to the traceability in
place.

Signed-off-by: Cédric Ollivier <cedric.ollivier@orange.com>