[CNTT] Networking scope updates
Addressing @karinesevilla comments, I have made changes to make sure that the text does not implicate that we will do everything within CNTT.  As CNTT agreed at the beginning, the approach is in the first place to harmonize existing standards.  If  not possible, we should identify a gap and identify which community should take care of filling the gap.  CNTT may be but not must be such an identified community.  It may be ETSI or any project under LNF etc