[RM Ch03] Update chapter03.md Network (#1535)
* Update chapter03.md

* Updates to 3.4 Network

Reflecting remarks from @pgoyal01 and @TFredberg:
- typos fixed 
- I agree with @TFredberg about sticking to the original "integration reference points", see his argument 
- actually I propose to drop "stack" and leave it as "different layers of the Cloud Infrastructure architecture", as we are speaking about architecture here, not deployment
- replace "on top' with the text " the cardinality of having multiple CaaS and IaaS instances executing on their own private virtual resources from the single shared HW Infrastructure Layer cannot be expressed into separate administrative domains.", which is a small simplification of the phrase proposed by @TFredberg 
My proposal is: @TFredberg @pgoyal01 and @ASawwaf : if you are OK with this, please approve immediately.  I would request  @rabi-abdel and @kedmison to approve to merge.  It could be good to have at least this "network manifesto" in Baldy.  In a sense we have beaten this small paragraph to death, which is good, because it will drive the rest of the work.  But we should not stop here !

* 3.4 Network

Link to Fig 3-1 added

* Update chapter03.md