[RA2 Ch04]: Changing musts to mays
However I believe that RA shold define hard requirements for components
visible for the software or the deployment artefacts of CNF-s and once
we made this decision in Prague I change the musts back to mays as there
are some concerns about this change in pr #1854.
The same changes will be submitted in a different pr.

Signed-off-by: Gergely Csatari <gergely.csatari@nokia.com>