Fix CNTT RC2 and RI1 versions
CNTT RC2 works well and is already in-use in Orange RFP [1].
-alpha is falsy misleading the endusers about its quality.

CNTT RI1 has never been executed in OPNFV since Jun 9, 2020.
The last run deploys OpenStack Ocata which is far from Train as
expected by CNTT RA1 (it should be noted that RI1 was alpha in Baldy
for the same OpenStack version.

[1] http://testresults.opnfv.org/functest/rfp/

Signed-off-by: Cédric Ollivier <cedric.ollivier@orange.com>