Remove definition of PVP (#2193)
PVP is one of the data-path topologies specified in TST009 for testing data-plane performance. The terminology section in this glossary should be expanded to be comprehensive for terminology in RM/RA/RI/RC ... OR ... IMO it is more helpful for a reader to have general terminology here () and very narrow terms defined or referenced in the relevant Anuket spec.