Add most of sigs in RA2 chapter 6 (#2400)
It remains sig-storage which asks for a deep dive on both features
and drivers. As starting point, the features list conforms with the
upstream K8s gates (kind).

It allows RA2 team to update the mandatory choices which RC2 will
conform. It's worth mentioning that a few related test could be still
skipped in RC2 (only for gce, disruptive, alpha, etc.).

Signed-off-by: Cédric Ollivier <cedric.ollivier@orange.com>

Co-authored-by: Cédric Ollivier <cedric.ollivier@orange.com>