Clean ONAP VNF Test Platform (VTP)
This obsolete chapters are not maintained and
continously breaks the gate.

Most of this chapter can be safely removed as
there is nothing more than publibicity about
VTP without neither requirements nor test cases.

Last not but least, it's not in the scope of the
next GSMA uplift including RA1+RC1.

Signed-off-by: Cédric Ollivier <cedric.ollivier@orange.com>