[gov] Remove empty vendor implementation documentation (#2967)
The documentation currently contains a vendor implementation section, which only
contains empty templates. Given that I am not aware of any ongoing or planned
attempts to detail the concept of a vendor implementation in Anuket, I propose
removing these templates. As a replacement, the term vendor implementation can
be explained in the glossary.

Please note: With the latest move to a newer version of Sphinx, these files do
not get rendered on read-the-docs anymore because these templates are still in
markdown. So, effectively, this section is already gone from our official
output.

Fixes #2966

Signed-off-by: Georg Kunz <georg.kunz@ericsson.com>