[docs] Fixing broken link to terms of reference
The link to the terms of references assumes that the pdf has been copied
to read-the-docs. This changes the link to instead point to the lastest
version of the terms of reference on GitHub. It should always be up-to-date
(i.e., the latest version) and not correspond to a released version of the
specs.

Fixes #3081

Signed-off-by: Georg Kunz <georg.kunz@ericsson.com>