[RM CH08] Additions of O-RAN alignment - Final
This PR describe O-RAN Cloud related parts on a very high level, its similarities and differences towards Anuket Infrastructure as well as a possible usage of Anuket as an under-cloud to O-RAN.
This is a clean updated version of PR #3345 & #3346 that was also debated as #3367. These are all closed and replaced by this PR.
Referenced figures in png and pptx are already uploaded since before. This procedure was decided on the RM meeting on May 31st.